refactored all scss in the application. each component now has its own scss
