+++ src/App.scss
@@ -1,5 +1,5 @@
-@import "variables";
-@import "mixins";
+@import "./src/styles/variables";
+@import "./src/styles/mixins";
 
 * {
   font-family: $text-font;
@@ -103,51 +103,7 @@
     "card-list"
     "footer-section";
 
-  .nav-bar {
-    //nav
-    //background: $complementary-color1;
-
-    /*divides navbar into two columns with a 1 to 3 ratio where nav-list is allowed 3/4 of horizontal space*/
-    display: grid;
-    grid-template-columns: 2fr 3fr;
-
-    /*adds a thin border separation between navbar & search box*/
-    border-bottom-style: solid;
-    border-bottom-color: $complementary-color2;
-    border-width: thin;
-
-    .nav-logo {
-      //header
-      //background:orange;
-
-      /*keeps logo in the middle of its container nav-logo*/
-      display: flex;
-      justify-content: center;
-    }
-    .nav-list {
-      //ul
-      //background: grey;
-      text-align: right;
-
-      /*keeps nav links item/s aligned to the right inside nav-list*/
-      display: flex;
-      align-items: center;
-      justify-content: flex-end;
-      margin-right: 30px; //<<adds a 30px margin to .nav-list container & fill middle gap/visual issue
-
-      .nav-link {
-        //use this class within .nav-list to automatically insert new link next to existing link
-        //a
-        //background: brown;
-        //color: $hyperlinks-color1;
-        display: inline; //keep nav links next to each other horizontally
-   
-        //padding: 10px;
-        margin-right: 20px; /*<<consider using mixin here to readjust margin-right in different viewports>>*/
-        font-size: 20px; //<<>>
-      }
-    }
-  }
+  
 
   // .header-section {
   //   //header
@@ -159,175 +115,6 @@
   //   }
   // }
 
-  #search-section {
-    //div
-    //background: $dark-background-color;
-  
-
-    /*Aligns search-bar & search-options inside a single column grid*/
-    display: grid;
-    height: 100%;
-    grid-template-columns: 1fr;
-    grid-template-rows: 2fr 1fr;
-    grid-template-areas:
-      "search-bar"
-      "search-options";
-
-      border-bottom-style: solid;
-      border-bottom-color: $complementary-color2;
-      border-width: thin;
-
-    .search-bar {
-      //div
-      //background-color: indigo;
-      //height: 50px;
-
-      min-width: 280px; //fixed min-width of 280px for search bar to prevent misalignment of search bar items
-      padding-top: 5px;
-      margin: auto;
-      
-      .input-field {
-        //input
-      }
-      .search-btn {
-        //button
-        //font-size: larger;
-        margin-left: 10px; //specify margin of 10px to keep a reasonable UI/UX standard
-        padding: 10px; //sets a padding of 10px to match search input field's padding
-      }
-      // .search-err-msg{ //div
-      // }
-    }
-    .search-options {
-      //background-color: blue;
-      margin: 0 auto; //brings search options closer to search bar
-      //margin-bottom: -30px; 
-      label {
-        //background: indianred;
-        margin: 0 5px; //adds space between options
-        color: $text-color2;
-      }
-    }
-  }
-
-  .card-list {
-    //div
-    //background: $complementary-color4;
-    overflow-y: auto; //enables vertical scrolling on card-list item & allows footer to remain sticky
-    display: grid; 
-    grid-template-columns: repeat(1, 1fr);
-
-    .card {
-      //div
-      //background: lightseagreen;
-      display: grid;
-      max-width: 380px; //set each card to a width of 430px to maintain consistency
-      //height: 220px; //set each card to a fixed height of 220px to maintain consitency 
-      margin: 20px auto; //applies a margin of 20px to each card for <480px viewport & centers in the middle of its container
-      /*divides card into two sections with the image on the left & card details on the right*/
-      grid-template-areas: "card-left card-right";
-      padding: 10px; 
-      border-bottom-style: solid;
-      border-bottom-color: $success-color;
-      border-width: thin;
-      
-
-      .card-left {
-        //background: palevioletred;
-        grid-area: card-left; //identifer for the left side of card & its img item
-        margin: 5px 15px; //applies a constent margin (left/right) to card items inside card-right in <480px
-
-        .card-img {
-          //img
-          //background: greenyellow;
-        }
-      }
-
-      .card-right {
-        //background: powderblue;
-        grid-area: card-right; //identifier for the right side of card & its items
-        display: grid;
-        margin: 5px 15px; //applies a consistent margin (left/right) to card items inside card-right in <480px 
-        color: $text-color2;
-
-        .card-title {
-          //h1
-          //background: cyan;  
-          font-size: 20px; //<<>>
-     
-        }
-        .card-subtitle {
-          //h2
-          //background: grey;
-          font-size: 15px; //<<>>
-        }
-        .card-authors {
-          //p
-          //background: orange;
-          font-size: 10px; //<<>>
-        }
-        .card-publisher {
-          //p
-          //background: lightcoral;
-          font-size: 10px; //<<>>
-        }
-        .card-btn {
-          //button
-          //background: purple;
-          margin: auto;
-        }
-      }
-    }
-  }
-
-  .footer-section {
-    //footer
-    //background: $dark-background-color;
-
-    /*aligns footer items in the middle of footer-section*/
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-
-    border-top-style: solid;
-    border-top-color: $complementary-color2;
-    border-width: thin;
-
-    //margin: auto;
-    //justify-content: center;
-    .footer-social {
-      //background: grey;
-
-      text-align: center; //keeps social media icons/links centered by default
-
-      margin-top: 5px; //<<>> matches margin space set on .footer-note
-
-      .footer-link {
-        //a
-        //background: green;
-
-       // color: $text-color1;
-
-        font-size: 10px; //<<>>
-        margin: 0 5px; //<<>> adds 5px margin between icons
-
-        //margin: auto;
-      }
-      .footer-link:hover {
-        //color: $hyperlinks-color2 !important;
-      }
-    }
-
-    .footer-note {
-      //p
-      //background: greenyellow;
-      color: $text-color2;
-      text-align: center;
-
-      font-size: 10px; //<<>>
-      margin: 5px 0; //<<>> matches 5px margin set on .footer-social's margin-top
-    }
-  }
 }
 
 // /* Extra small (xs) 320px — 480px: Mobile devices*/
