+++ src/components/Card/Card.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 
-import Modal from "./Modal";
+import Modal from "../Modal/Modal";
 
 // functional component responsible to render one individual Card aka Volume (Book, Magazine or Newspaper)
 export default function Card({ book, query }) {
@@ -29,7 +29,7 @@ export default function Card({ book, query }) {
   return (
 <>
 <div className="modal-overlay" style={style} onClick={toggleModal}></div>
-    <div className="card" style={{zIndex: "100"}}>
+    <div className="card">
       <div className="card-left">
         <img
           className="card-image"
