+++ src/components/Search/Search.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SearchBar from "./SearchBar";
+import SearchBar from "../SearchBar/SearchBar";
 
 export default function Search({ handleSubmit, error, queryHistory }) {
   return (
